加入发送接收结构体，舵机控制

+++ SYSTEM/usart.h
@@ -3,7 +3,7 @@
 #include "stdio.h"	
 #include "sys.h" 
  	
-extern u16 USART_RX_STA;         		//接收状态标记	
+extern u8 USART_RX_STA;         		//接收状态标记	
 
 void UART_GPS_Init(u32 bound);
 void UART_24G_Init(u32 bound);
